+++ torchmetrics/regression/__init__.py
@@ -17,9 +17,9 @@ from torchmetrics.regression.mean_absolute_error import MeanAbsoluteError  # noq
 from torchmetrics.regression.mean_absolute_percentage_error import MeanAbsolutePercentageError  # noqa: F401
 from torchmetrics.regression.mean_squared_error import MeanSquaredError  # noqa: F401
 from torchmetrics.regression.mean_squared_log_error import MeanSquaredLogError  # noqa: F401
-from torchmetrics.regression.pearson import PearsonCorrcoef  # noqa: F401
+from torchmetrics.regression.pearson import PearsonCorrCoef, PearsonCorrcoef  # noqa: F401
 from torchmetrics.regression.r2 import R2Score  # noqa: F401
-from torchmetrics.regression.spearman import SpearmanCorrcoef  # noqa: F401
+from torchmetrics.regression.spearman import SpearmanCorrCoef, SpearmanCorrcoef  # noqa: F401
 from torchmetrics.regression.symmetric_mean_absolute_percentage_error import (  # noqa: F401
     SymmetricMeanAbsolutePercentageError,
 )